repositories: add support for empty CWD in pkg_* phases.
EAPI=8 requires the CWD to be set to an empty directory at the start of
each pkg_* phase.

Older EAPIs have no provisions regarding this, which means that,
strictly speaking, they also do not disallow such usage.

This implementation will hence try to use an empty CWD opportunistically
on Gentoo, even if the EAPI does not require it. If anything fails,
older EAPIs will continue normally, while on EAPI=8+, failing to change
to an empty CWD will lead to a hard error.

The requirement is rather strict, which also means that just creating an
empty directory and then sequentially using that to execute the pkg_*
phases without purging/recreating it would violate it if a phase
actually creates files or directories within it.

Thus, make sure to recreate it for every phase.

Exherbo code is not affected by this change. On this platform, the
phases will execute the same way they always did.