ruby: add BDEPEND and IUSE_EFFECTIVE to test_each_metadata function.
BDEPEND should be self-explanatory (since we now test for that variable,
too), but IUSE_EFFECTIVE needs a further comment.

Normally, IUSE_EFFECTIVE is not exported to the ebuild environment,
since it's calculated from IUSE and other variables (like IUSE_EXPAND).
However, for EAPIs 5+, the package manager must additionally save the
calculated value, due to the profile-defined IUSE injection feature for
each installed package. Profiles can change, but for installed packages,
the whole state is to be saved, which means that such external factors
must be explicitly saved.