tests: refactor empty pkg_* cwd EAPI=8 test.
Instead of copying a lot of code to each phase, use a custom function
and call that in each phase.

Additionally, try to make the CWD dirty by creating a file, to see if
subsequent pkg_* phases actually see empty CWDs as well. Accept
failures, since PMS specifies that the empty CWD need not be writeable.