tests: rewrite econf EAPI=7 test in pure sh.
Using constructs like "echo foo | grep ..." opens the door to a lot of
potential pitfalls, since the echo builtin is a real beast. User input
must be specially sanitized to make sure that things work correctly,
which is difficult. Using shell-only features and the printf builtin is
much safer in most circumstances.

Additionally, the old code technically checked for 'build', 'target' and
'with-sysroot' only, but not actually if passed as arguments (i.e.,
missing the dash-dash specifier) and their actual values for simplicity
with grep. By switching to a for loop, we can just check for the real
deal.

Also, just for weird reasons, make CBUILD and CHOST more fruity.